requirements.txt
PyMuPDF
pillow
requests
openai
python-dotenv
discord.py
